requirements.txt
langchain==0.1.19
langchain-community==0.0.38
langchain-openai==0.1.7
openai>=1.10.0,<2.0.0

faiss-cpu==1.8.0.post1
tiktoken==0.7.0
python-dotenv==1.0.0
numpy>=1.26.0,<2.0.0

fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
setuptools>=68.0.0
wheel>=0.41.0
